@@ -1,8 +1,8 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
     defaultConfig {
         applicationId "com.stfalcon.chatkit.sample"
         minSdkVersion 14
@@ -21,28 +21,34 @@ android {
             abortOnError false
         }
     }
+
+    compileOptions {
+        sourceCompatibility 1.8
+        targetCompatibility 1.8
+    }
 }
 
 ext {
-    supportVersion = '27.1.1'
+    androidxVersion = '1.0.2'
+    materialVersion = '1.1.0-alpha03'
     picassoVersion = '2.5.2'
     circleImageViewVersion = '2.2.0'
     shapeImageViewVersion = '0.9.3'
-    circleindicatorVersion = '1.2.2@aar'
+    circleindicatorVersion = '2.1.0'
 }
 
 dependencies {
     implementation project(':chatkit')
 
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
-        exclude group: 'com.android.support', module: 'support-annotations'
-    })
     testImplementation 'junit:junit:4.12'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.0'
+    androidTestImplementation 'androidx.test:core:1.1.0'
+    androidTestImplementation 'androidx.test:runner:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
 
-    implementation "com.android.support:appcompat-v7:$supportVersion"
-    implementation "com.android.support:cardview-v7:$supportVersion"
-    implementation "com.android.support:design:$supportVersion"
+    implementation "androidx.appcompat:appcompat:$androidxVersion"
+    implementation "com.google.android.material:material:$materialVersion"
+    implementation 'androidx.cardview:cardview:1.0.0'
 
     //Picasso
     implementation "com.squareup.picasso:picasso:$picassoVersion"
@@ -53,5 +59,4 @@ dependencies {
 
     //Utils
     implementation "me.relex:circleindicator:$circleindicatorVersion"
-
 }
