@@ -2,7 +2,6 @@ package com.stfalcon.chatkit.sample.features.main.adapter;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -10,6 +9,8 @@ import android.widget.Button;
 import android.widget.TextView;
 
 import com.stfalcon.chatkit.sample.R;
+
+import androidx.fragment.app.Fragment;
 
 /*
  * Created by troy379 on 11.04.17.
