@@ -1,11 +1,10 @@
 package com.stfalcon.chatkit.sample.features.main.adapter;
 
 import android.content.Context;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentStatePagerAdapter;
-
 import com.stfalcon.chatkit.sample.R;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentStatePagerAdapter;
 
 /*
  * Created by troy379 on 11.04.17.
